Probably finalized CPersistent Onject. Tried both CArrayContainer and CMongoContainer with it and it works.
Now I need to implement the NewObject static method, should probably create a subclass of CContainer that looks for the class name in the object and instantiates it. Another strategy is to create an object that consumes CContainer derived classes and takes care of instantiating the right class. Then I need to integrate this with the cursor concept.

Need to place the SerialiseObject static method & Co. maybe the CWrapper class?