Finalised CContainer, CArrayContainer and CPersistentObject.
I am currently working on CMongoContainer on the _Commit() method, finished the _PrepareCommit() method and added parent _Prepare and _Finish methods, so remember to delete them, except _PrepareCommit().

Need to finalize CMongoContainer and decide whether to provide a copy of the object to the persistence methods or leave it as-is.