Finalised CQuery and CWrapper families. Tested wrapper and it works.
What is pending?

There is still the issue of data types in queries: you still need to repeat the data type both in the clause and in the data, we could get rid of it, but it means a lot of work, so we leave it like this today.

Queries are essentially built with arrays, we can make the process easier by adding methods to the query classes.

How can I send a query via POST? That needs to be added as a static method to the wrapper ancestor.