Working on an object to handle saving nodes and edges to Mongo. Finished COntologyNodeIndex class, need to develop the edge counterpart.
I updated index management in COntologyNode and must still test it.
I still need to develop the edge counterpart and test it.

Next I need to update all places that use the Lucene index, this could be very teavvy.

Spotted a problem in the MATCH web service: need to check again how queries can be provided: cannot manage to provide an AND list of clauses.

Once all the above will be done we shall be again operative.